Update pdu.js
Since each pduHeaderParam is 4 bytes/octets, the "offset" is equal to the total length of the pduHeadParams.length * 4. Using the basis is maintenance friendly. I noticed this while trying to leverage your code base to build a module for CUAP USSD protocol for USSDC that only support TCP/IP CUAP protocol.